Workers/tests/submissions/Opera/interfaces/WorkerUtils/importScripts/002.html

<!--
var x = 0;
try {
  importScripts('data:text/javascript,x=1','http://foo bar');
} catch(e) {
  x += 2
}
postMessage(x);
/*
-->
<!doctype html>
<title>importScripts resolving urls</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  var worker = new Worker('#');
  worker.onmessage = this.step_func(function(e) {
    assert_equals(e.data, 2);
    this.done();
  });
});
</script>
<!--
*/
//-->
